WebServer-RTS-v2/requirements.txt
pyyaml==6.0.1
paho-mqtt==1.6.1
firebase-admin==6.4.0
